+++ src/main/java/view/ClientWindow.java
@@ -21,6 +21,7 @@ public class ClientWindow extends JFrame {
     private GameState gameState;
     private Font dozerFont;
     private JButton startButton;
+    private JLabel statusLabel;
 
     /**
      * Instantiates a new Client window.
@@ -130,6 +131,7 @@ public class ClientWindow extends JFrame {
             waitingPanel.add(buttonPanel, gbc);
 
             setContentPane(waitingPanel);
+            addStatusLabel();
             revalidate();
             repaint();
         } catch (Exception e) {
@@ -161,8 +163,8 @@ public class ClientWindow extends JFrame {
                 playerListModel.addElement(player);
                 TypingPlayer newPlayer = new TypingPlayer(player);
                 gameState.addPlayer(newPlayer);
-
             }
+            updateLobbyStatus();
         });
     }
 
@@ -175,22 +177,38 @@ public class ClientWindow extends JFrame {
         JOptionPane.showMessageDialog(this, playerName + " has left the game.", "Player Left", JOptionPane.INFORMATION_MESSAGE);
     }
 
+    public void addStatusLabel () {
+        statusLabel = new JLabel();
+        statusLabel.setFont(dozerFont.deriveFont(Font.PLAIN, 20));
+        statusLabel.setHorizontalAlignment(JLabel.CENTER);
+        statusLabel.setForeground(Color.WHITE);
+
+        GridBagConstraints gbc = new GridBagConstraints();
+        gbc.gridwidth = GridBagConstraints.REMAINDER;
+        gbc.anchor = GridBagConstraints.CENTER;
+        gbc.fill = GridBagConstraints.HORIZONTAL;
+        gbc.insets = new Insets(10, 10, 10, 10);
+
+        JPanel waitingPanel = (JPanel) getContentPane();
+        waitingPanel.add(statusLabel, gbc);
+    }
+
     /**
      * Show lobby full button.
      */
-    public void showLobbyFullButton() {
-        JButton lobbyFullButton = create3DButton("Lobby full, Start the Game");
-        lobbyFullButton.addActionListener(e -> {
-            if (clientController.isHost(playerName)) {
-                clientController.startGame(playerName);
-            } else {
-                JOptionPane.showMessageDialog(this, "You are not the host. Only the host can start the game.", "Access Denied", JOptionPane.ERROR_MESSAGE);
-            }
-        });
-        JPanel waitingPanel = (JPanel) getContentPane();
-        waitingPanel.add(lobbyFullButton, BorderLayout.SOUTH);
+    public void updateLobbyStatus() {
+        if (isLobbyFull()) {
+            statusLabel.setText("Lobby full, please start the game");
+        } else {
+            statusLabel.setText("");
+        }
         revalidate();
         repaint();
+    }
+
+
+    private boolean isLobbyFull() {
+        return playerListModel.size() == 6;
     }
 
 /*    public void startGame() {