+++ CHANGELOG.md
@@ -13,9 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # NETWORK:  
 ### Assignees : _Yili Li and Yuanyuan Qu_
 
+## 2024-07-11
+### Fixed
+- lobby full: show label "lobby full start the game", when one player left, label disappear
+
 ## 2024-07-10
 ### Fixed
-- join game by entering ip-adress
+- join game by entering ip-address
 
 ## 2024-07-09
 ### Added